chore: manually fix the stubs
I don't know how to do this without manually editing the stubs.
It was already like this before, so I suppose that someone already did
the edit before.

I tried to update everything and use the standard protobuf to manage
includes and path but it doesn't seems to work. Object exchanged with
etcd daemon failed to be decrypted here.