build: removing generated files from gconf
Tox is responsible for delivering these files using the build target.
Moreover, if the working directory is dirty, ToX automatically increment
the current version to a devX.<hash> instead of a given potential tag,
to ensure that the delivered tarball do match the tag associated to a
gconf state.
This reason impose such a removing for proper delivery